Bugfixes
it builds, it runs, it does what its supposed to do

+++ Core/Src/sensor/communication.h
@@ -16,7 +16,7 @@
 
 /// The D6T-32L-01A (that i was given measures in a 32x32 = 1024 pattern
 #define TEMP_VALUE_COUNT 1024
-#define ADDRESS 0x14
+#define ADDRESS (uint8_t)0x14
 #define CONNECTION_TRYS 10
